DesafioEmbarca/processCsvAndSaveToDb/requirements.txt
pandas
boto3
psycopg2-binary
sqlalchemy
awslambdaric